+++ tests/testthat/test-read-stata.R
@@ -22,3 +22,9 @@ test_that("Stata %td (date) and %tc (datetime) read into expected classes", {
   expect_is(df$vdate, "Date")
   expect_is(df$vdatetime, "POSIXct")
 })
+
+
+test_that("Old %d format read into Date class", {
+  df <- read_stata(test_path("datetime-d.dta"))
+  expect_equal(df$date, as.Date("2015-11-02"))
+})